services/stock_info/requirements.txt
FinMind
Flask
pandas
numpy
numpyencoder
scikit-learn
joblib
pyaml